update for ui with reload list function
Saw several requests on reloading the whole list without the page without an answer how to do it. As I had the same issue I added this. It works but I'm just a JS beginner.
Would be nice if someone could review this if it is valid and makes sense.