Refactored Angular module.
Renamed fileupload directive to fileUpload (file-upload in templates)
for consistency with the service and controllers.
Preview and Progress directives now use "file-upload-" as prefix as a
sort of namespace.
The formatFileSizeFilter now accepts suffix and prefix for the units.
This allows better localizing for languages which put the unit label
first.