Allow aborting of jqXHR through promise.abort
The jqXHR object contained in the promise is overwritten by a promise object lacking an abort method via chaining `_onSend().then`.

Instead, jqXHR should be retained as a jqXHR even though the promise methods can be attached to it.

Try using:

    $("input").fileUpload().fileupload("send" {files: element}).abort()

to get an error that the `abort` method does not exist in this context.