useSubmit option
Currently, fileupload only works either a) when the file is selected, or b) manually via API. However, often we want to submit the form via Ajax. There are complications here - for example, we want to allow for optional file entry, where the form is still submitted even if no file is selected. The only current way around that is a complex and non-DRY method where we copy all the Ajax settings we passed into the fileupload plugin and have a separate submit listener that replicates the functionality only if no file has been selected.

This change adds a ``useSubmit`` option. In this case, selecting a file won't fire the send method; instead, it is *always* fired when the form is submitted, regardless of whether a file has been added or not. All existing callbacks should still work exactly as expected, including ``_onAdd``.