finfo mimetype upload checking
Uses finfo to check actual contents of uploaded file for mime type and
not just the file extention.  If user renames .exe to .jpg this method
will detect that.

If user doesn't have finfo, wants to use regex or has an older version
of script - it will fall back to original method gracefully.