Bring back support for CommonJS
Browserify, which follows CommonJS module conventions, parses the AST
for `require` calls to determine the dependencies of a project.

jQuery File Upload used to support this; it was removed in
e2cda462610c776a1f7856692c98b1baab02231a when a new version of
jquery.ui.widget.js was vendored.