Initialization on input element for use in jQuery-UI fails
Extension for use in JQuery-UI had an issue when fileupload was
initialized on the input type=file element directly rather than on the
form.
In fact use of `this.element´ had to be replaced by a higher level in
DOM so we can .find() it always.
So I replaced the element by the fileInput.